wip refactor to allow cli mode clean
